test(app.test.js): 경주 진행 로직 테스트 추가
