+++ __tests__/App.test.js
@@ -1,0 +1,44 @@
+import App from '../src/App.js';
+import { MissionUtils } from '@woowacourse/mission-utils';
+
+const mockQuestions = inputs => {
+  MissionUtils.Console.readLineAsync = jest.fn();
+  MissionUtils.Console.readLineAsync.mockImplementation(() => {
+    const input = inputs.shift();
+    return Promise.resolve(input);
+  });
+};
+
+const mockRandoms = numbers => {
+  MissionUtils.Random.pickNumberInRange = jest.fn();
+  numbers.reduce(
+    (acc, number) => acc.mockReturnValueOnce(number),
+    MissionUtils.Random.pickNumberInRange,
+  );
+};
+
+const getLogSpy = () => {
+  const logSpy = jest.spyOn(MissionUtils.Console, 'print');
+  logSpy.mockClear();
+  return logSpy;
+};
+
+describe('App 통합 테스트', () => {
+  test('경주 실행', async () => {
+    const inputs = ['pobi,woni', '1'];
+    const logs = ['pobi : -', 'woni : ', '최종 우승자 : pobi'];
+    const logSpy = getLogSpy();
+
+    mockQuestions(inputs);
+    mockRandoms([4, 3]);
+
+    // when
+    const app = new App();
+    await app.run();
+
+    // then
+    logs.forEach(log => {
+      expect(logSpy).toHaveBeenCalledWith(expect.stringContaining(log));
+    });
+  });
+});
